update: modify interface
